main.py
import numpy as np
import streamlit as st
import ai_edge_litert.interpreter as litert
from PIL import Image, ImageOps

def classify(image, model, class_names):
    # Convert image to (224, 224)
    image = ImageOps.fit(image, (224, 224), Image.Resampling.LANCZOS)

    # Convert image to numpy array
    image_array = np.asarray(image)

    # Normalize image
    normalized_image_array = image_array.astype(np.float32) / 127.5 - 1

    # Set model input
    data = np.ndarray(shape=(1, 224, 224, 3), dtype=np.float32)
    data[0] = normalized_image_array

    input_details = model.get_input_details()
    output_details = model.get_output_details()   
    model.set_tensor(input_details[0]['index'], data)
    model.invoke()

    # Make prediction
    prediction = model.get_tensor(output_details[0]['index'])
    index = np.argmax(prediction[0])
    class_name = class_names[index]
    confidence_score = prediction[0][index]

    return class_name, format(confidence_score * 100, '.2f')

# Set title
st.title('Pneumonia Classification')

# Set header
st.header('Please upload a chest X-Ray image')

# Upload file
file = st.file_uploader('', type=['jpeg', 'jpg', 'png'])

# Load classifier
model_path = './model/model_unquant.tflite'
model = litert.Interpreter(model_path=model_path)
model.allocate_tensors()

# Load class names
with open('./model/labels.txt', 'r') as f:
    class_names = [class_[:-1].split(' ')[1] for class_ in f.readlines()]

# Display image
if file:
    image = Image.open(file).convert('RGB')
    st.image(image)

    # Classify image
    class_name, conf_score = classify(image, model, class_names)

    # Write classification
    st.write(f'## {class_name}')
    st.write(f'### Score: {conf_score}%')