requirements.txt
Pillow
numpy
streamlit
ai-edge-litert